Make sure that Wine Python works
This is because I spent a few days trying to work out what was going
wrong (especially as I was able to do it by hand successfully) when it
kept coming up with errors regarding overly long path lengths.

After thinking it was due to Ansible somehow breaking everything, I
eventually realised I was half-right: due to Ansible being written in
Python, to get it working properly on NixOS I need to run it in a
nix-shell where the PYTHONPATH is set for you... but this contains
many long paths which don't work (and aren't needed) inside of the
Wine environment.

As such, you need to make sure you specify an empty PYTHONPATH for
wine.

I thought about creating a new role just for this but it seemed a tad
overkill just for one tasks with many other things to test.